requirements.txt
#
# This file is autogenerated by pip-compile with Python 3.11
# by the following command:
#
#    pip-compile
#
asgiref==3.8.1
    # via django
black==24.4.2
    # via -r requirements.in
build==1.2.1
    # via pip-tools
click==8.1.7
    # via
    #   black
    #   pip-tools
django==5.0.6
    # via
    #   -r requirements.in
    #   django-extensions
    #   djangorestframework
django-extensions==3.2.3
    # via -r requirements.in
djangorestframework==3.15.2
    # via -r requirements.in
inflection==0.5.1
    # via -r requirements.in
iniconfig==2.0.0
    # via pytest
isort==5.13.2
    # via -r requirements.in
mypy-extensions==1.0.0
    # via black
packaging==24.1
    # via
    #   black
    #   build
    #   pytest
pathspec==0.12.1
    # via black
pip-tools==7.4.1
    # via -r requirements.in
platformdirs==4.2.2
    # via black
pluggy==1.5.0
    # via pytest
pyproject-hooks==1.1.0
    # via
    #   build
    #   pip-tools
pytest==8.2.2
    # via
    #   -r requirements.in
    #   pytest-django
pytest-django==4.8.0
    # via -r requirements.in
pyyaml==6.0.1
    # via -r requirements.in
sqlparse==0.5.0
    # via django
uritemplate==4.1.1
    # via -r requirements.in
wheel==0.43.0
    # via pip-tools

# The following packages are considered to be unsafe in a requirements file:
# pip
# setuptools
